requirements/requirements.txt
Django
djangorestframework
markdown
django-filter
djoser
djangorestframework_simplejwt
psycopg2-binary
drf-yasg
drf-nested-routers
django-cors-headers